Rename more site sections and add redirects I wanted to tackle a few more site sections before we merged this. This also adds a set of redirects that should hopefully cover all the changes we're making, both now and in the future.
