+++ uber/templates/emails/mits/mits_waiver.txt
@@ -1,6 +1,6 @@
 {{ team.salutation }},
 
-Congratulations on being accepted into the {{ c.EVENT_NAME }} MITS Showcase! We ask that all teams presenting at {{ c.EVENT_NAME }} sign a waiver and policy form. You can review the waiver and sign it at {{ c.URL_BASE }}/mits_applications/waiver?id={{ team.id }}.
+Congratulations on being accepted into the {{ c.EVENT_NAME }} MITS Showcase! We ask that all teams presenting at {{ c.EVENT_NAME }} sign a waiver and policy form. You can review the waiver and sign it at {{ c.URL_BASE }}/mits/waiver?id={{ team.id }}.
 
 Signing the waiver before the event makes things easier for everyone, so please do so ASAP! In the meantime, let us know if you have any questions!
