+++ uber/templates/mits_admin/team.html
@@ -4,7 +4,7 @@
 <h2 class="btn-toolbar">
     {{ team.name }}
     <a href="delete_team?id={{ team.id }}" class="btn btn-danger pull-right">Delete Application</a>
-    <a href="../mits_applications/continue_app?id={{ team.id }}" class="btn btn-primary pull-right">Edit Application</a>
+    <a href="../mits/continue_app?id={{ team.id }}" class="btn btn-primary pull-right">Edit Application</a>
 </h2>
 
 <div class="form-horizontal">
