+++ uber/templates/mits_applications/team.html
@@ -17,7 +17,7 @@
 {% else %}
     {% if team.is_new %}
         <h2>MAGFest Indie Tabletop Showcase Application</h2>
-        Have you already filled this out? If so, click <a href="../mits_applications/check_if_applied">here</a>. <br/><br/>
+        Have you already filled this out? If so, click <a href="../mits/check_if_applied">here</a>. <br/><br/>
     {% else %}
         <h2>Edit Your Team Information</h2>
     {% endif %}
