+++ uber/templates/registration/review.html
@@ -8,6 +8,6 @@
 {# When we switch from Django to Jinja2, this can be injected into the main registration form #}
 {% block extra %}
     {% if 'hotel eligib' in attendee.for_review|lower and not attendee.hotel_eligible and c.HAS_HOTEL_ADMIN_ACCESS %}
-        ++++++ <a href="../hotel/mark_hotel_eligible?id={{ attendee.id }}">Admin: Force mark as hotel eligible</a>
+        ++++++ <a href="../hotel_admin/mark_hotel_eligible?id={{ attendee.id }}">Admin: Force mark as hotel eligible</a>
     {% endif %}
 {% endblock extra %}
